requirements.txt
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
docutils==0.14
entrypoints==0.3
gunicorn==20.0.4
html5lib==1.0.1
jsonschema==3.0.1
Markdown==3.1.1
requests==2.21.0
tables==3.5.1
